Move the HubSpotPromise type to the types/http file
