+++ api/fileManager.ts
@@ -1,7 +1,7 @@
 import fs from 'fs';
 import path from 'path';
-import { http, HubSpotPromise } from '../http';
-import { FormData } from '../types/Http';
+import { http } from '../http';
+import { FormData, HubSpotPromise } from '../types/Http';
 import {
   FetchStatResponse,
   FetchFilesResponse,
