+++ api/fileTransport.ts
@@ -1,7 +1,8 @@
 import fs from 'fs';
 import path from 'path';
 import { getCwd } from '../lib/path';
-import { http, HubSpotPromise } from '../http';
+import { http } from '../http';
+import { HubSpotPromise } from '../types/Http';
 const HUBFILES_API_PATH = '/file-transport/v1/hubfiles';
 
 export function createSchemaFromHubFile(
