+++ api/projects.ts
@@ -1,6 +1,6 @@
-import { http, HubSpotPromise } from '../http';
+import { http } from '../http';
 import fs from 'fs';
-import { FormData, QueryParams } from '../types/Http';
+import { FormData, HubSpotPromise, QueryParams } from '../types/Http';
 import {
   Project,
   FetchProjectResponse,
