+++ api/sandboxSync.ts
@@ -1,10 +1,11 @@
-import { http, HubSpotPromise } from '../http';
+import { http } from '../http';
 import {
   InitiateSyncResponse,
   FetchTypesResponse,
   TaskRequestData,
 } from '../types/Sandbox';
 import { SANDBOX_TIMEOUT } from '../constants/api';
+import { HubSpotPromise } from '../types/Http';
 const SANDBOXES_SYNC_API_PATH = 'sandboxes-sync/v1';
 
 export async function initiateSync(
