+++ api/validateHubl.ts
@@ -1,5 +1,6 @@
-import { http, HubSpotPromise } from '../http';
+import { http } from '../http';
 import { Validation, HublValidationOptions } from '../types/HublValidation';
+import { HubSpotPromise } from '../types/Http';
 
 const HUBL_VALIDATE_API_PATH = 'cos-rendering/v1/internal/validate';
